Move pipeline_generator to pipegen dir in root, add go.mod file
