--- buildkite/tools/pipeline_generator/update/main.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"fmt"
-)
-
-func main() {
-	// TODO(fweikert): implement
-	fmt.Printf("Not yet implemented.")
-}
