+++ go.mod
@@ -1,0 +1,13 @@
+module github.com/bazelbuild/continuous-integration
+
+go 1.12
+
+require (
+	github.com/buildkite/go-buildkite v2.2.0+incompatible
+	github.com/cenkalti/backoff v2.1.1+incompatible // indirect
+	github.com/fweikert/go-buildkite v2.1.1-0.20190131135720-1a18fa668fae+incompatible
+	github.com/google/go-querystring v1.0.0 // indirect
+	github.com/machinebox/graphql v0.2.2
+	github.com/pkg/errors v0.8.1 // indirect
+	gopkg.in/yaml.v2 v2.2.2
+)
