+++ pipegen/README.md
@@ -1,0 +1,3 @@
+go get gopkg.in/yaml.v2
+go get github.com/buildkite/go-buildkite/buildkite
+go get github.com/machinebox/graphql
