+++ pipegen/update/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"fmt"
+)
+
+func main() {
+	// TODO(fweikert): implement
+	fmt.Printf("Not yet implemented.")
+}
